#1 refactor: CalendarCategory 위치 변경
